api cleaning

+++ app/Http/Controllers/API/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Api\UserApiResource;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\GameResource;
 use App\Models\Media;
@@ -16,39 +17,17 @@ class UserController extends Controller
 {
     public function show($username)
     {
-        $user = User::select(['id', 'username', 'display_name', 'media_id'])
-            ->with('games', function ($query) {
-                return $query->paginate(12);
-            })
+        $user = User::select(['id', 'username', 'display_name', 'media_id', 'avatar'])
             ->withCount('games')
             ->where('username', $username)
             ->firstOrFail();
 
-        $games = $user->games;
-        unset($user->games);
-        $user->games = GameResource::collection($games);
-        unset($user->following);
-
-        $total_pages = ceil($user->games_count / 12);
-        $next_page = 2;
-        $next_page_url = null;
-        if (request()->query('page'))
-            $next_page = request()->query('page') + 1;
-
-        if ($next_page > $total_pages)
-            $next_page = null;
-
-        if ($next_page)
-            $next_page_url = route('api.user') . '/' . $username . '?page=' . $next_page;
-
-        $user->last_page = $total_pages;
-        $user->next_page_url = $next_page_url;
+        $user->games = $user->games()->with('status')->paginate(12);
 
         return response()->api(
-            data: $user,
+            data: UserApiResource::parse($user),
             message: $user->username . __(" Profile")
         );
-
     }
 
     public function update(Request $request)
@@ -80,7 +59,7 @@ class UserController extends Controller
         $user->save();
 
         return response()->api(
-            data: $user,
+            data: UserApiResource::parse($user),
             message: __('Information updated successfully')
         );
     }
@@ -121,7 +100,7 @@ class UserController extends Controller
         $user->save();
 
         return response()->api(
-            data: $user,
+            data: UserApiResource::parse($user),
             message: __('Password changed successfully')
         );
     }
@@ -149,7 +128,7 @@ class UserController extends Controller
         $user->save();
 
         return response()->api(
-            data: $user,
+            data: UserApiResource::parse($user),
             message: __('Email changed successfully')
         );
     }
@@ -171,9 +150,13 @@ class UserController extends Controller
             );
         }
 
-        $file = public_path('media/images/users/avatars/'.$user->avatar);
-        if (file_exists($file))
-            unlink($file);
+        if($user->avatar)
+        {
+            $file = public_path('media/images/users/avatars/'.$user->avatar);
+            if (file_exists($file))
+                unlink($file);
+        }
+
 
         $avatar = Media::uploadFile(file: $request->file('avatar'), path: "/images/users/avatars/", size: [200, 200]);
 
@@ -181,7 +164,7 @@ class UserController extends Controller
         $user->save();
 
         return response()->api(
-            data: $user,
+            data: UserApiResource::parse($user),
             message: __('Avatar changed successfully')
         );
     }
